Converted seconds offset to integer to eliminate rounding errors.

+++ adafruit_ntp.py
@@ -56,7 +56,7 @@ class NTP:
         self._server = server
         self._port = port
         self._packet = bytearray(48)
-        self._tz_offset = tz_offset * 60 * 60
+        self._tz_offset = int(tz_offset * 60 * 60)
         self._socket_timeout = socket_timeout
 
         # This is our estimated start time for the monotonic clock. We adjust it based on the ntp
